backend/requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
httpx==0.24.1
httpcore==0.17.3
certifi==2024.2.2